+++ src/abacus-backoffice/jest.config.js
@@ -1,0 +1,8 @@
+// @flow strict
+
+const path = require('path');
+
+module.exports = {
+  rootDir: __dirname,
+  setupFilesAfterEnv: [(path.join(__dirname, 'jest-setup.js') /*: string */)],
+};