+++ src/sx-design/src/Layout/__tests__/LayoutGrid.test.js
@@ -27,6 +27,6 @@ it.each`
       <Placeholder width={100} height={100} />
     </LayoutGrid>,
   );
-  expect(getByTestId('layout_grid_test_id')).toBeDefined();
+  expect(getByTestId('layout_grid_test_id')).toBeInTheDocument();
   expect(getByTestId('layout_grid_test_id')).toHaveStyle({ gap: expectedGap });
 });